refactor: codebloc and markdownimage move to component folder
