+++ components/CodeBlock.js
@@ -1,0 +1,12 @@
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import style from 'react-syntax-highlighter/dist/cjs/styles/prism/dracula';
+
+function CodeBlock({ language, value }) {
+  return (
+    <SyntaxHighlighter style={style} language={language}>
+      {value}
+    </SyntaxHighlighter>
+  );
+}
+
+export default CodeBlock;
